@@ -3,5 +3,8 @@
 const cueRuntime = require('..');
 
 describe('cue-runtime', () => {
-    it('needs tests');
+  test('module mode preamble', () => {
+    const code = 'function() {}';
+    expect(code).toMatch('function() {}')
+  })
 });
